Started work on 2D renderer

+++ Hanabi/src/Engine/Renderer/OpenGL/OpenGLVertexArray.cpp
@@ -53,7 +53,7 @@ namespace Hanabi
 		for (const auto& element : layout)
 		{
 			glEnableVertexAttribArray(m_VertexBufferIndex);
-			glVertexAttribPointer(m_VertexBufferIndex, element.GetComponentCount(), ShaderDataTypeToOpenGLBaseType(element.Type), element.Normalized ? GL_TRUE : GL_FALSE, layout.GetStride(), (const void*)(intptr_t)element.Offset);
+			glVertexAttribPointer(m_VertexBufferIndex, element.GetComponentCount(), ShaderDataTypeToOpenGLBaseType(element.Type), element.Normalized ? GL_TRUE : GL_FALSE, layout.GetStride(), (const void*)element.Offset);
 			m_VertexBufferIndex++;
 		}
 		m_VertexBuffers.push_back(vertexBuffer);
